--- test/sourcemaps/samples/script-and-style/test.js
@@ -1,9 +1,0 @@
-import { assert_mapped } from '../../helpers';
-
-export function test({ input, preprocessed }) {
-	assert_mapped({
-		code: '<h1>Hello world!</h1>',
-		input: input.locate,
-		preprocessed
-	});
-}
