+++ test/sourcemaps/samples/sourcemap-concat/test.js
@@ -2,7 +2,7 @@ import { assert_mapped } from '../../helpers';
 
 export function test({ input, preprocessed }) {
 	assert_mapped({
-    code: 'Target',
+		code: 'Target',
 		input: input.locate,
 		preprocessed
 	});
