+++ test/sourcemaps/samples/sourcemap-offsets/input.svelte
@@ -1,0 +1,9 @@
+<script>
+	// This block is here to offset style block
+</script>
+
+<style>
+	div { --component-var: 2px; }
+</style>
+
+<div><span>Text</span></div>
